move addHeaders to api-client/display and stub out new fetchApiDisplay()
