+++ src/api-client/display.cpp
@@ -1,0 +1,53 @@
+#include <api-client/display.h>
+#include <ArduinoLog.h>
+#include <HTTPClient.h>
+#include <trmnl_log.h>
+#include <WiFiClientSecure.h>
+#include <config.h>
+#include <api_response_parsing.h>
+#include <http_client.h>
+
+void addHeaders(HTTPClient &https, ApiDisplayInputs &inputs)
+{
+  Log_info("Added headers:\n\r"
+           "ID: %s\n\r"
+           "Special function: %d\n\r"
+           "Access-Token: %s\n\r"
+           "Refresh_Rate: %s\n\r"
+           "Battery-Voltage: %s\n\r"
+           "FW-Version: %s\r\n"
+           "RSSI: %s\r\n",
+           __FILE__, __LINE__,
+           inputs.macAddress.c_str(),
+           inputs.specialFunction,
+           inputs.apiKey.c_str(),
+           String(inputs.refreshRate).c_str(),
+           String(inputs.batteryVoltage).c_str(),
+           inputs.firmwareVersion.c_str(),
+           String(inputs.rssi));
+
+  https.addHeader("ID", inputs.macAddress);
+  https.addHeader("Access-Token", inputs.apiKey);
+  https.addHeader("Refresh-Rate", String(inputs.refreshRate));
+  https.addHeader("Battery-Voltage", String(inputs.batteryVoltage));
+  https.addHeader("FW-Version", inputs.firmwareVersion);
+  https.addHeader("RSSI", String(inputs.rssi));
+  https.addHeader("Width", String(inputs.displayWidth));
+  https.addHeader("Height", String(inputs.displayHeight));
+
+  if (inputs.specialFunction != SF_NONE)
+  {
+    Log.info("%s [%d]: Add special function: true (%d)\r\n", __FILE__, __LINE__, inputs.specialFunction);
+    https.addHeader("special_function", "true");
+  }
+}
+
+ApiDisplayResult fetchApiDisplay(ApiDisplayInputs &apiDisplayInputs)
+{
+  ApiDisplayResponse response = {};
+  return ApiDisplayResult{
+      .error = https_request_err_e::HTTPS_NO_ERR,
+      .response = response,
+      .error_detail = "",
+  };
+};
